refactor: Initial refactoring for REGAPIC transcoding
Isolating the transcoder from RestMethod will allow for simpler
data-driven testing and a clean way of handling additional bindings
(as the RestMethod will then just have multiple transcoders).

This commit is relatively minimal - there are further aspects of
refactoring which will be easier to reason about after this is in.

This commit does not introduce any tests, but these will be added in
later commits when we get to the core part of transcoding.
